feat(builder): retry stuck bundles, fix maxFeePerGas
If a bundle goes for two blocks without being mined, send a replacement
transaction with the same nonce but higher gas fees to replace it. The
replacement transaction is a newly generated bundle with stricter gas
fee requirements to account for the new higher fees.

There are a few edge cases to watch out for in the implementation:

* We submit a replacement transaction, but the earlier one ends up
  getting mined.
* Our account's nonce increases from a separate transaction not related
  to the bundle transactions, invalidating any outstanding transactions.
* While we initially have ops to send, after increasing the gas fees
  there are no longer any ops that meet the fee requirements.

Introduce several new enums to represent the various failure cases.

Further, fix handling of `maxFeePerGas` on bundle transactions, as well
as the requirement for `maxFeePerGas` on user ops, checked in both the
prechecker and the proposer. We now send our transactions using a
`maxFeePerGas` calculated from the current base fee increased by 12.5%,
and place equivalent requirements on user ops.

As part of the new focus on making sure that `maxFeePerGas` is correct,
introduce a new struct, `GasFees`, containing the two fee fields, and
refactor code throughout the codebase to use it.

This is a breaking change, as previously accepted user ops will now be
rejected because their `maxFeePerGas` is too low.

+++ src/cli/pool.rs
@@ -54,7 +54,7 @@ pub struct PoolArgs {
         env = "POOL_MIN_REPLACEMENT_FEE_INCREASE_PERCENTAGE",
         default_value = "10"
     )]
-    pub min_replacement_fee_increase_percentage: usize,
+    pub min_replacement_fee_increase_percentage: u64,
 
     /// ETH Node HTTP polling interval in milliseconds
     /// (only used if node_http is set)
